make transformations non-commutative, new transforms
Transformations now depend on the order in which they're used. This is closer to the students' intuitions from math (or ones that they will study later). It is alos more efficient and less complicated in the code. Also added some new transforms and exposed the advanced linear algebra for students which might benefit from it, and for us to use. Note that most students should not try to understand this way of transforming, the regular transforms are intended to be used by most users.